+++ ttmp/2025/12/28/002-ADD-TITLE-DESCRIPTION--add-node-title-description/changelog.md
@@ -3,4 +3,5 @@
 ## 2025-12-28
 
 - Initial workspace created
+- Added design spec for node title/description + instantiation note, with verification scripts guidance